Transaction id is a u32

+++ src/model/id/transaction_id.rs
@@ -1,4 +1,4 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Copy, Clone, Deserialize, Serialize, PartialEq, Eq, Hash)]
-pub struct TransactionId(pub u16);
+pub struct TransactionId(pub u32);
